requirements.txt
vizeval>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
rich>=13.0.0 